+++ src/components/Sidebar/Project/DeleteProject.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function DeleteProject() {
+  return (
+    <div>DeleteProject</div>
+  )
+}
