+++ src/components/Sidebar/Project/index.ts
@@ -1,0 +1,1 @@
+export { default as SidebarProject } from './Project';
