--- src/components/provider/index.ts
@@ -1,2 +1,0 @@
-export * from './TaskFilterProvider';
-export * from './ProjectsProvider';
