--- src/context/hook/index.ts
@@ -1,2 +1,0 @@
-export * from "./useTaskFilterContext";
-export * from "./useProjectsContext";
