--- src/context/hook/useProjectsContext.ts
@@ -1,4 +1,0 @@
-import {useContext} from 'react';
-import ProjectsContext from '../ProjectsContext';
-
-export const useProjectsContext = () => useContext(ProjectsContext);
