--- src/context/hook/useTaskFilterContext.ts
@@ -1,4 +1,0 @@
-import {useContext} from 'react';
-import TaskFilterContext from '../TaskFilterContext';
-
-export const useTaskFilterContext = () => useContext(TaskFilterContext);
