--- src/feature/sidebar/index.ts
@@ -1,3 +1,0 @@
-export { default as AddTask } from './AddTask';
-export { default as Filter } from './Filter';
-export { default as ProjectList } from './ProjectList';
