--- src/utils/helpers.ts
@@ -1,8 +1,0 @@
-import {Project} from '@/types';
-
-// ProjectIdからProjectItemを取得.
-export const getProject = (
-  projects: Project[],
-  projectId: string,
-): Project | undefined =>
-  projects.find((project) => project.projectId === projectId);
