+++ src/views/ProjectView.tsx
@@ -1,5 +1,5 @@
 import {Tasks} from '@/components/feature/task/Tasks';
-import {Header} from '@/components/layout';
+import {Header} from '@/components/Layout';
 import {useTasks} from '@/hook/useTasks';
 import {useParams} from 'react-router-dom';
 
@@ -10,7 +10,6 @@ type Params = {
 export default function ProjectView(): React.ReactNode {
   const params = useParams<Params>();
   const projectId = params.projectId ?? null;
-
   const tasks = useTasks({projectId, isArchive: false});
